+++ tests/test_parse_clinvar.py
@@ -96,10 +96,10 @@ def test_parse_clinvar_snv_grch37(input_clinvar_snv, output_single, output_multi
     # Check subsequent
     lines_single = load_tsv(output_single)
     assert len(lines_single) == 1
-    assert lines_single[0][:6] == ["GRCh37", "13", "95228658", "T", "C", "95228658"]
+    assert lines_single[0][:7] == ["GRCh37", "13", "95228658", "95228658", "1311", "T", "C"]
     assert lines_single[0][-1:] == ["0"]
     # Check clinical significance
-    assert lines_single[0][19:24] == ['0', '1', '0', '0', '0']
+    assert lines_single[0][19:24] == ["0", "1", "0", "0", "0"]
     lines_multi = load_tsv(output_multi)
     assert lines_multi == []
 
@@ -114,10 +114,10 @@ def test_parse_clinvar_snv_grch38(input_clinvar_snv, output_single, output_multi
     # Check subsequent
     lines_single = load_tsv(output_single)
     assert len(lines_single) == 1
-    assert lines_single[0][:6] == ["GRCh38", "chr13", "94576404", "T", "C", "94576404"]
+    assert lines_single[0][:7] == ["GRCh38", "chr13", "94576404", "94576404", "1306", "T", "C"]
     assert lines_single[0][-1:] == ["0"]
     # Check clinical significance
-    assert lines_single[0][19:24] == ['0', '1', '0', '0', '0']
+    assert lines_single[0][19:24] == ["0", "1", "0", "0", "0"]
     lines_multi = load_tsv(output_multi)
     assert lines_multi == []
 
@@ -139,9 +139,9 @@ def test_parse_clinvar_two_measure_sets_grch37(
     assert lines_single == []
     lines_multi = load_tsv(output_multi)
     assert len(lines_multi) == 2
-    assert lines_multi[0][:6] == ["GRCh37", "7", "103276772", "C", "T", "103276772"]
+    assert lines_multi[0][:7] == ["GRCh37", "7", "103276772", "103276772", "1372", "C", "T"]
     assert lines_multi[0][-1:] == ["1"]
-    assert lines_multi[1][:6] == ["GRCh37", "7", "103132416", "A", "C", "103132416"]
+    assert lines_multi[1][:7] == ["GRCh37", "7", "103132416", "103132416", "1371", "A", "C"]
     assert lines_multi[1][-1:] == ["1"]
 
 
@@ -162,7 +162,7 @@ def test_parse_clinvar_two_measure_sets_grch38(
     assert lines_single == []
     lines_multi = load_tsv(output_multi)
     assert len(lines_multi) == 2
-    assert lines_multi[0][:6] == ["GRCh38", "chr7", "103636325", "C", "T", "103636325"]
+    assert lines_multi[0][:7] == ["GRCh38", "chr7", "103636325", "103636325", "1375", "C", "T"]
     assert lines_multi[0][-1:] == ["1"]
-    assert lines_multi[1][:6] == ["GRCh38", "chr7", "103491969", "A", "C", "103491969"]
+    assert lines_multi[1][:7] == ["GRCh38", "chr7", "103491969", "103491969", "1374", "A", "C"]
     assert lines_multi[1][-1:] == ["1"]